added member data to chip8 m_lastPressedKey so that instruction ldVxK works properly now. Modified the keyboard to array of bools because realized I don't need to know the order of the pressed keys

+++ src/base64/base64.h
@@ -7,7 +7,7 @@
 #include <string>
 #include <cstdint>
 
-std::string base64_encode(const char* buf, unsigned int bufLen);
+std::string base64_encode(const char* buf, size_t bufLen);
 
 std::vector<uint8_t> base64_decode(const char* encoded_string, size_t in_len);
